chore(plugins/api): make `api.endpoints` option array instead of object and refactor plugin routes and middleware into folders
