+++ lib/plugins/api/resources/mod.ts
@@ -1,5 +1,9 @@
 import type { ULID } from "../../../deps/ulid.ts";
 
+export * from "./custom.ts";
+export * from "./denokv.ts";
+export * from "./http.ts";
+
 export type Method = "find" | "get" | "create" | "update" | "patch" | "remove";
 
 export type ResourceOptions = {
